requirements.txt
pandas
scikit-learn
tensorflow
mlflow
imbalanced-learn
fastapi
uvicorn[standard]
pyyaml
pytest